backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8
drf-yasg==1.21.7
django-filter==23.5
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
django-allauth==0.60.1
requests==2.31.0
pyotp==2.9.0
qrcode==7.4.2
stripe==7.11.0
cryptography==41.0.7 